requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.0
certifi==2022.12.7
charset-normalizer==2.1.1
DonationAlertsAPI==2.0.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
urllib3==1.26.13
websocket-client==1.4.2
websockets==10.4
yarl==1.8.2